Extract module for console options
